fix: move interface files to src file and possible fix for transient db test failures
https://github.com/FiddlersCode/leopold/issues/89
