+++ tests/unit/leopoldMongoClientWrapper.spec.ts
@@ -1,9 +1,9 @@
 import {MongoMemoryServer} from 'mongodb-memory-server';
 import * as mongodb from 'mongodb';
 import {LeopoldMongoClientWrapper, MongoCollections} from "../../src/mongo/leopoldMongoClientWrapper";
-import {Concert} from "../../src/mongo/concert"
+import {Concert} from "../../src/concert"
 import {makeTestConcert} from "../testHelpers"
-import {Gig} from "../../src/mongo/gig";
+import {Gig} from "../../src/gig";
 
 
 // @ts-ignore
@@ -18,43 +18,17 @@ describe('Leopold Mongo Client Wrapper', () => {
     let gigsCollection: any;
     let mongod: MongoMemoryServer;
 
-    beforeAll(async () => {
+    beforeAll(async (done) => {
         mongod = new MongoMemoryServer();
         const mongoUri = await mongod.getUri();
         leopoldMongoClientWrapper = new LeopoldMongoClientWrapper(mongoUri);
         await leopoldMongoClientWrapper.connectToDB(testDB);
-        gigsCollection = leopoldMongoClientWrapper.dbConnection.collection(MongoCollections.GIGS);
+        gigsCollection = leopoldMongoClientWrapper.db.collection(MongoCollections.GIGS);
+        done();
     });
 
-    describe('DB Connections', () => {
-        it('ss100.T10: open database connection', () => {
-            const expectedLogMessage: string = `${testDB} connection open.`;
-            expect(global.console.log).toHaveBeenCalledWith(expectedLogMessage)
-        });
-    });
-    describe('DB insertions', () => {
-
-        it('ss100.T10: add a gig to the db', (done) => {
-            const dressCode = "ss100.T10";
-            const concertToAdd: Concert | Gig = makeTestConcert(dressCode);
-
-            leopoldMongoClientWrapper.addGig(concertToAdd);
-
-            const callback = (doc) => {
-                const expectedLogMessage = `Successfully created entry id: ${doc._id}`;
-                const expectedLogMessage1 = `${testDB} connection open.`;
-                expect(global.console.log).toHaveBeenCalledWith(expectedLogMessage1);
-                expect(global.console.log).toHaveBeenCalledWith(expectedLogMessage);
-                expect(doc).toEqual(concertToAdd);
-                done();
-            };
-
-            gigsCollection.findOne({dressCode}, (err, doc) => {
-                if (err) console.log(err);
-                callback(doc);
-            });
-        });
-        it('ss100.T15: add multiple gigs to the db', async () => {
+    describe('ss100: DB insertions', () => {
+        it('ss100.T20: add multiple gigs to the db', async () => {
             const dressCode = "ss100.T15";
             const numberOfGigs = 2;
             const gigsToAdd: any[] = [];
@@ -68,28 +42,7 @@ describe('Leopold Mongo Client Wrapper', () => {
             expect(gigsAdded.length).toEqual(numberOfGigs);
 
         });
-        it('ss100.T20 log an error if there is a problem adding one gig', (done) => {
-            const dressCode = "ss100.T20";
-            const gigToAdd: Concert | Gig = makeTestConcert(dressCode);
-
-            leopoldMongoClientWrapper.addGig(gigToAdd);
-            leopoldMongoClientWrapper.addGig(gigToAdd);
-
-            const callback = (doc) => {
-                gigsCollection.insertOne({_id: doc._id}, (err) => {
-                    if (err) console.log(err);
-                });
-                const expectedLogMessage = new mongodb.MongoError(`E11000 duplicate key error dup key: { : ObjectId('${doc._id}') }`);
-                expect(global.console.error).toHaveBeenCalledWith(expectedLogMessage);
-                done();
-            };
-
-            gigsCollection.findOne({dressCode}, (err, doc) => {
-                if (err) console.log(err);
-                callback(doc);
-            });
-        });
-        it('ss100.T25 log an error if there is a problem adding multiple gigs', (done) => {
+        it('ss100.T20 log an error if there is a problem adding multiple gigs', (done) => {
             const dressCode = "ss100.T25";
             const gigToAdd: Concert | Gig = makeTestConcert(dressCode);
 
@@ -113,7 +66,7 @@ describe('Leopold Mongo Client Wrapper', () => {
     });
 
     afterAll(async (done) => {
-        await leopoldMongoClientWrapper.dbConnection.dropCollection(MongoCollections.GIGS);
+        await leopoldMongoClientWrapper.db.dropCollection(MongoCollections.GIGS);
         await leopoldMongoClientWrapper.dbConnection.close();
         leopoldMongoClientWrapper = null;
         mongod = null;
